+++ tests/websocket/close-dropped-server/e2e.test.ts
@@ -1,6 +1,6 @@
 import SleepySocketClient from 'sleepy-socket'
-import { CloseCode } from 'sleepy-serv'
 import { describe, mock, test, expect } from 'bun:test'
+import { KickedCloseSignal } from './src/utils'
 
 import {
   Fmt,
@@ -20,7 +20,7 @@ const OPTS_RECONNECT = {
 }
 
 describe('handler', () => {
-  test('when NO custom reason provided AND reconnect enabled', async () => {
+  test('when dropped provided AND reconnect enabled', async () => {
     const server = await createServer(import.meta.dirname)
     const reqClient = createClient(server)
     const handlers = [mock(), mock(), mock()]
@@ -37,7 +37,7 @@ describe('handler', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/handler-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/handler-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -51,7 +51,12 @@ describe('handler', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -62,13 +67,13 @@ describe('handler', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])
   })
 
-  test('when NO custom reason provided AND reconnect disabled', async () => {
+  test('when dropped provided AND reconnect disabled', async () => {
     const server = await createServer(import.meta.dirname)
     const reqClient = createClient(server)
     const handlers = [mock(), mock(), mock()]
@@ -85,7 +90,7 @@ describe('handler', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/handler-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/handler-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -98,7 +103,12 @@ describe('handler', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -109,101 +119,7 @@ describe('handler', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect enabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const reqClient = createClient(server)
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        OPTS_RECONNECT,
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/handler-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect disabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const reqClient = createClient(server)
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        { reconnect: false },
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/handler-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])
@@ -211,7 +127,7 @@ describe('handler', () => {
 })
 
 describe('middleware', () => {
-  test('when NO custom reason provided AND reconnect enabled', async () => {
+  test('when dropped provided AND reconnect enabled', async () => {
     const server = await createServer(import.meta.dirname)
     const reqClient = createClient(server)
     const handlers = [mock(), mock(), mock()]
@@ -228,7 +144,7 @@ describe('middleware', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/middleware-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/middleware-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -242,7 +158,12 @@ describe('middleware', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -253,13 +174,13 @@ describe('middleware', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])
   })
 
-  test('when NO custom reason provided AND reconnect disabled', async () => {
+  test('when dropped provided AND reconnect disabled', async () => {
     const server = await createServer(import.meta.dirname)
     const reqClient = createClient(server)
     const handlers = [mock(), mock(), mock()]
@@ -276,7 +197,7 @@ describe('middleware', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/middleware-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/middleware-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -289,7 +210,12 @@ describe('middleware', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -300,101 +226,7 @@ describe('middleware', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect enabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const reqClient = createClient(server)
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        OPTS_RECONNECT,
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/middleware-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect disabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const reqClient = createClient(server)
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        { reconnect: false },
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/middleware-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])
@@ -402,7 +234,7 @@ describe('middleware', () => {
 })
 
 describe('app', () => {
-  test('when NO custom reason provided AND reconnect enabled', async () => {
+  test('when dropped provided AND reconnect enabled', async () => {
     const server = await createServer(import.meta.dirname)
     const adminPort = await getAdminPort(server)
     const reqClient = createClient({ port: adminPort })
@@ -420,7 +252,7 @@ describe('app', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/app-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/app-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -434,7 +266,12 @@ describe('app', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -445,13 +282,13 @@ describe('app', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])
   })
 
-  test('when NO custom reason provided AND reconnect disabled', async () => {
+  test('when dropped provided AND reconnect disabled', async () => {
     const server = await createServer(import.meta.dirname)
     const adminPort = await getAdminPort(server)
     const reqClient = createClient({ port: adminPort })
@@ -469,7 +306,7 @@ describe('app', () => {
       client.on('close', handlers[index])
     })
 
-    await reqClient.post(`/app-trigger-drop-default`, Fmt.Text, {
+    await reqClient.post(`/app-trigger-drop`, Fmt.Text, {
       body: JSON.stringify({
         clientId: clients[1].id,
       }),
@@ -482,7 +319,12 @@ describe('app', () => {
     expect(clients[2].isConnected).toBe(true)
     expect(handlers[0]).not.toHaveBeenCalled()
     expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+    expect(handlers[1]).toHaveBeenCalledWith({
+      code: KickedCloseSignal.code,
+      reason: KickedCloseSignal.reason,
+    })
+
     expect(handlers[2]).not.toHaveBeenCalled()
 
     await clients[0].close()
@@ -493,103 +335,7 @@ describe('app', () => {
     const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
     expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:ok`, /* CHECK: reason = "dropped"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect enabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const adminPort = await getAdminPort(server)
-    const reqClient = createClient({ port: adminPort })
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        OPTS_RECONNECT,
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/app-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
-      `CLOSE:${clients[0].id}:ok`,
-      `CLOSE:${clients[2].id}:ok`,
-    ])
-  })
-
-  test('when CUSTOM reason provided AND reconnect disabled', async () => {
-    const server = await createServer(import.meta.dirname)
-    const adminPort = await getAdminPort(server)
-    const reqClient = createClient({ port: adminPort })
-    const handlers = [mock(), mock(), mock()]
-
-    const clients = await Promise.all(
-      handlers.map(() => SleepySocketClient.open(
-        'localhost',
-        server.port,
-        { reconnect: false },
-      )),
-    )
-
-    clients.forEach((client, index) => {
-      client.on('close', handlers[index])
-    })
-
-    await reqClient.post(`/app-trigger-drop-custom`, Fmt.Text, {
-      body: JSON.stringify({
-        clientId: clients[1].id,
-      }),
-    })
-
-    await waitFor(() => !clients[1].isConnected)
-    await waitForCloseCount(server, 1)
-
-    expect(clients[0].isConnected).toBe(true)
-    expect(clients[2].isConnected).toBe(true)
-    expect(handlers[0]).not.toHaveBeenCalled()
-    expect(handlers[1]).toHaveBeenCalledOnce()
-    expect(handlers[1]).toHaveBeenCalledWith({ code: 3999 })
-    expect(handlers[2]).not.toHaveBeenCalled()
-
-    await clients[0].close()
-    await clients[2].close()
-    await waitForCloseCount(server, 3)
-    await server.kill()
-
-    const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
-
-    expect(closeLines).toStrictEqual([
-      `CLOSE:${clients[1].id}:dropped`, /* CHECK: reason = "player_kicked"? */
+      `CLOSE:${clients[1].id}:kicked`, /* CHECK: reason = "dropped"? */
       `CLOSE:${clients[0].id}:ok`,
       `CLOSE:${clients[2].id}:ok`,
     ])