+++ tests/websocket/close-dropped-server/src/api/middleware-trigger-drop-default/post.ts
@@ -1,3 +1,0 @@
-export default function handler (): Response {
-  return new Response('', { status: 204 })
-}