+++ tests/websocket/close-ok/e2e.test.ts
@@ -1,5 +1,5 @@
 import SleepySocketClient from 'sleepy-socket'
-import { CloseCode } from 'sleepy-serv'
+import { InternalCloseSignal } from 'sleepy-serv'
 import { mock, test, expect } from 'bun:test'
 import { createServer, waitFor } from '../../helpers'
 
@@ -23,7 +23,12 @@ test('when closed from client AND reconnect enabled', async () => {
   const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
   expect(handler).toHaveBeenCalledOnce()
-  expect(handler).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+  expect(handler).toHaveBeenCalledWith({
+    code: InternalCloseSignal.Ok.code,
+    reason: InternalCloseSignal.Ok.reason,
+  })
+
   expect(closeLines).toHaveLength(1)
   expect(closeLines[0]).toBe(`CLOSE:${client.id}:ok`)
 })
@@ -47,7 +52,12 @@ test('when closed from client AND reconnect disabled', async () => {
   const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
   expect(handler).toHaveBeenCalledOnce()
-  expect(handler).toHaveBeenCalledWith({ code: CloseCode.Ok })
+
+  expect(handler).toHaveBeenCalledWith({
+    code: InternalCloseSignal.Ok.code,
+    reason: InternalCloseSignal.Ok.reason,
+  })
+
   expect(closeLines).toHaveLength(1)
   expect(closeLines[0]).toBe(`CLOSE:${client.id}:ok`)
 })