+++ tests/websocket/close-reaped/e2e.test.ts
@@ -1,5 +1,5 @@
 import SleepySocketClient from 'sleepy-socket'
-import { CloseCode } from 'sleepy-serv'
+import { InternalCloseSignal } from 'sleepy-serv'
 import { mock, test, expect } from 'bun:test'
 import { createServer, wait, waitFor } from '../../helpers'
 
@@ -30,8 +30,17 @@ test('when reaped AND reconnect enabled', async () => {
   const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
   expect(handler).toHaveBeenCalledTimes(2)
-  expect(handler).toHaveBeenNthCalledWith(1, { code: CloseCode.Reaped })
-  expect(handler).toHaveBeenNthCalledWith(2, { code: CloseCode.Ok })
+
+  expect(handler).toHaveBeenNthCalledWith(1, {
+    code: InternalCloseSignal.Reaped.code,
+    reason: InternalCloseSignal.Reaped.reason,
+  })
+
+  expect(handler).toHaveBeenNthCalledWith(2, {
+    code: InternalCloseSignal.Ok.code,
+    reason: InternalCloseSignal.Ok.reason,
+  })
+
   expect(closeLines).toHaveLength(2)
   expect(closeLines[0]).toBe(`CLOSE:${client.id}:reaped`)
   expect(closeLines[1]).toBe(`CLOSE:${client.id}:ok`)
@@ -55,7 +64,12 @@ test('when reaped AND reconnect disabled', async () => {
   const closeLines = server.output.filter(l => l.startsWith('CLOSE:'))
 
   expect(handler).toHaveBeenCalledOnce()
-  expect(handler).toHaveBeenCalledWith({ code: CloseCode.Reaped })
+
+  expect(handler).toHaveBeenCalledWith({
+    code: InternalCloseSignal.Reaped.code,
+    reason: InternalCloseSignal.Reaped.reason,
+  })
+
   expect(closeLines).toHaveLength(1)
   expect(closeLines[0]).toBe(`CLOSE:${client.id}:reaped`)
 })